add auth + refactor
