--- backend/internal/github/app/github.go
@@ -1,10 +1,0 @@
-package github
-
-import (
-	gh "github.com/google/go-github/v52/github"
-)
-
-type GitHubApp interface {
-	Ping() (string, error)
-	ListRepos() ([]*gh.Repository, error)
-}
