+++ backend/internal/github/client/api/api.go
@@ -1,138 +1,97 @@
-package api
+package github
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"net/http"
-	"net/url"
-	"strings"
 
 	"github.com/CamPlume1/khoury-classroom/internal/config"
+	gh "github.com/CamPlume1/khoury-classroom/internal/github"
+	"github.com/CamPlume1/khoury-classroom/internal/models"
 	"github.com/google/go-github/v52/github"
 	"golang.org/x/oauth2"
 )
 
 type API struct {
-	client *github.Client
+	gh.GitHubBase
+	appConfig oauth2.Config
+	client    *github.Client
 }
 
 // https://github.com/login/oauth/authorize?client_id=Ov23liswNMlwZUn1hnmS&scope=repo,read:org,classroom&allow_signup=false
 
 func New(cfg *config.GitHubClient, code string) (*API, error) {
-	// OAuth token URL for exchanging code for access token
-	tokenURL := "https://github.com/login/oauth/access_token"
-
-	// Prepare the request
-	data := url.Values{}
-	data.Set("client_id", cfg.ClientID)
-	data.Set("client_secret", cfg.ClientSecret)
-	data.Set("code", code)
-
-	// Create the request
-	req, err := http.NewRequest("POST", tokenURL, strings.NewReader(data.Encode()))
+	// Exchange the authorization code for an access token
+	oAuth2Config := cfg.OAuthConfig()
+	token, err := oAuth2Config.Exchange(context.Background(), code)
 	if err != nil {
-		return nil, fmt.Errorf("error creating token request: %v", err)
+		return nil, fmt.Errorf("error exchanging code for token: %v", err)
 	}
-
-	// Set headers
-	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-	req.Header.Set("Accept", "application/json")
-
-	// Make the request
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("error making token request: %v", err)
-	}
-	defer resp.Body.Close()
-
-	// Decode response & extract the access token
-	var tokenResponse struct {
-		AccessToken string `json:"access_token"`
-		TokenType   string `json:"token_type"`
-		Scope       string `json:"scope"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&tokenResponse); err != nil {
-		return nil, fmt.Errorf("error decoding token response: %v", err)
-	}
-
-	// Create an OAuth2 token with the access token
-	token := &oauth2.Token{AccessToken: tokenResponse.AccessToken}
 
 	// Create an OAuth2 HTTP client
-	httpClient := oauth2.NewClient(context.Background(), oauth2.StaticTokenSource(token))
+	httpClient := oAuth2Config.Client(context.Background(), token)
 
 	// Create the GitHub client
 	githubClient := github.NewClient(httpClient)
 
 	return &API{
-		client: githubClient,
+		client:    githubClient,
+		appConfig: *oAuth2Config,
 	}, nil
 }
 
-func (api *API) ListRepositories(ctx context.Context) ([]*github.Repository, error) {
-	repos, _, err := api.client.Repositories.ListByOrg(ctx, "NUSpecialProjects", nil)
+// func New(cfg *config.GitHubClient, code string) (*API, error) {
+// 	// OAuth token URL for exchanging code for access token
+// 	tokenURL := "https://github.com/login/oauth/access_token"
 
-	if err != nil {
-		return repos, fmt.Errorf("Error fetching repositories: %v", err)
-	}
+// 	// Prepare the request
+// 	data := url.Values{}
+// 	data.Set("client_id", cfg.ClientID)
+// 	data.Set("client_secret", cfg.ClientSecret)
+// 	data.Set("code", code)
 
-	return repos, nil
-}
+// 	// Create the request
+// 	req, err := http.NewRequest("POST", tokenURL, strings.NewReader(data.Encode()))
+// 	if err != nil {
+// 		return nil, fmt.Errorf("error creating token request: %v", err)
+// 	}
 
-func (api *API) ListCommits(ctx context.Context, owner string, repo string, opts *github.CommitsListOptions) ([]*github.RepositoryCommit, error) {
-	commits, _, err := api.client.Repositories.ListCommits(ctx, owner, repo, opts)
+// 	// Set headers
+// 	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
+// 	req.Header.Set("Accept", "application/json")
 
-	return commits, err
-}
+// 	// Make the request
+// 	client := &http.Client{}
+// 	resp, err := client.Do(req)
+// 	if err != nil {
+// 		return nil, fmt.Errorf("error making token request: %v", err)
+// 	}
+// 	defer resp.Body.Close()
 
-func (api *API) GetBranch(ctx context.Context, owner_name string, repo_name string, branch_name string) (*github.Branch, error) {
-	branch, _, err := api.client.Repositories.GetBranch(ctx, owner_name, repo_name, branch_name, false)
+// 	// Decode response & extract the access token
+// 	var tokenResponse struct {
+// 		AccessToken string `json:"access_token"`
+// 		TokenType   string `json:"token_type"`
+// 		Scope       string `json:"scope"`
+// 	}
+// 	if err := json.NewDecoder(resp.Body).Decode(&tokenResponse); err != nil {
+// 		return nil, fmt.Errorf("error decoding token response: %v", err)
+// 	}
 
-	return branch, err
-}
+// 	// Create an OAuth2 token with the access token
+// 	token := &oauth2.Token{AccessToken: tokenResponse.AccessToken}
 
-type ClassroomAcceptedAssignment struct {
-	ID          int64               `json:"id"`
-	Submitted   bool                `json:"submitted"`
-	Passing     bool                `json:"passing"`
-	CommitCount int                 `json:"commit_count"`
-	Grade       string              `json:"grade"`
-	Students    []*github.User      `json:"students"`   // Using GitHub User type for students
-	Repository  *github.Repository  `json:"repository"` // Using GitHub Repository type
-	Assignment  ClassroomAssignment `json:"assignment"`
-}
+// 	// Create an OAuth2 HTTP client
+// 	httpClient := oauth2.NewClient(context.Background(), oauth2.StaticTokenSource(token))
 
-type ClassroomAssignment struct {
-	ID                          int64     `json:"id"`
-	PublicRepo                  bool      `json:"public_repo"`
-	Title                       string    `json:"title"`
-	Type                        string    `json:"type"`
-	InviteLink                  string    `json:"invite_link"`
-	InvitationsEnabled          bool      `json:"invitations_enabled"`
-	Slug                        string    `json:"slug"`
-	StudentsAreRepoAdmins       bool      `json:"students_are_repo_admins"`
-	FeedbackPullRequestsEnabled bool      `json:"feedback_pull_requests_enabled"`
-	MaxTeams                    *int      `json:"max_teams,omitempty"`   // Nullable int
-	MaxMembers                  *int      `json:"max_members,omitempty"` // Nullable int
-	Editor                      string    `json:"editor"`
-	Accepted                    int       `json:"accepted"`
-	Submitted                   int       `json:"submitted"`
-	Passing                     int       `json:"passing"`
-	Language                    string    `json:"language"`
-	Deadline                    *string   `json:"deadline,omitempty"` // Nullable datetime
-	Classroom                   Classroom `json:"classroom"`
-}
+// 	// Create the GitHub client
+// 	githubClient := github.NewClient(httpClient)
 
-type Classroom struct {
-	ID       int64  `json:"id"`
-	Name     string `json:"name"`
-	Archived bool   `json:"archived"`
-	URL      string `json:"url"`
-}
+// 	return &API{
+// 		client: githubClient,
+// 	}, nil
+// }
 
-func (api *API) ListClassrooms(ctx context.Context) ([]Classroom, error) {
+func (api *API) ListClassrooms(ctx context.Context) ([]models.Classroom, error) {
 	// Construct the URL for the classrooms endpoint
 	endpoint := "https://api.github.com/classrooms"
 
@@ -143,7 +102,7 @@ func (api *API) ListClassrooms(ctx context.Context) ([]Classroom, error) {
 	}
 
 	// Response container for classrooms
-	var classrooms []Classroom
+	var classrooms []models.Classroom
 
 	// Make the API call
 	_, err = api.client.Do(ctx, req, &classrooms)
@@ -154,7 +113,7 @@ func (api *API) ListClassrooms(ctx context.Context) ([]Classroom, error) {
 	return classrooms, nil
 }
 
-func (api *API) ListAssignmentsForClassroom(ctx context.Context, classroomID int64) ([]ClassroomAssignment, error) {
+func (api *API) ListAssignmentsForClassroom(ctx context.Context, classroomID int64) ([]models.ClassroomAssignment, error) {
 	// Construct the URL for the list assignments endpoint
 	endpoint := fmt.Sprintf("/classrooms/%d/assignments", classroomID)
 
@@ -165,7 +124,7 @@ func (api *API) ListAssignmentsForClassroom(ctx context.Context, classroomID int
 	}
 
 	// Response container for assignments
-	var assignments []ClassroomAssignment
+	var assignments []models.ClassroomAssignment
 
 	// Make the API call
 	_, err = api.client.Do(ctx, req, &assignments)
@@ -176,7 +135,7 @@ func (api *API) ListAssignmentsForClassroom(ctx context.Context, classroomID int
 	return assignments, nil
 }
 
-func (api *API) GetAcceptedAssignments(ctx context.Context, assignmentID int64) ([]ClassroomAcceptedAssignment, error) {
+func (api *API) GetAcceptedAssignments(ctx context.Context, assignmentID int64) ([]models.ClassroomAcceptedAssignment, error) {
 	// Construct the URL for the assignment endpoint
 	endpoint := fmt.Sprintf("/assignments/%d/accepted_assignments", assignmentID)
 
@@ -187,7 +146,7 @@ func (api *API) GetAcceptedAssignments(ctx context.Context, assignmentID int64) 
 	}
 
 	// Response container for accepted assignments
-	var acceptedAssignments []ClassroomAcceptedAssignment
+	var acceptedAssignments []models.ClassroomAcceptedAssignment
 
 	// Make the API call
 	_, err = api.client.Do(ctx, req, &acceptedAssignments)
