+++ backend/internal/github/github.go
@@ -1,0 +1,34 @@
+package github
+
+import (
+	"context"
+
+	// "github.com/CamPlume1/khoury-classroom/internal/config"
+	"github.com/CamPlume1/khoury-classroom/internal/models"
+	"github.com/google/go-github/v52/github"
+)
+
+type GitHubApp interface {
+	GitHubBase
+	// App only functionalities
+}
+
+type GitHubClient interface {
+	GitHubBase
+	ListClassrooms(ctx context.Context) ([]models.Classroom, error)
+	ListAssignmentsForClassroom(ctx context.Context, classroomID int64) ([]models.ClassroomAssignment, error)
+	GetAcceptedAssignments(ctx context.Context, assignmentID int64) ([]models.ClassroomAcceptedAssignment, error)
+}
+
+type GitHubBase interface {
+	Ping(ctx context.Context) (string, error)
+	ListRepositories(ctx context.Context) ([]*github.Repository, error)
+	ListCommits(ctx context.Context, owner string, repo string, opts *github.CommitsListOptions) ([]*github.RepositoryCommit, error)
+	GetBranch(ctx context.Context, owner_name string, repo_name string, branch_name string) (*github.Branch, error)
+	CreateBranch(ctx context.Context, owner string, repo string, baseBranch string, newBranchName string) error
+	CreatePullRequest(ctx context.Context, owner string, repo string, baseBranch string, headBranch string, title string, body string) (*github.PullRequest, error)
+	CreateInlinePRComment(ctx context.Context, owner string, repo string, pullNumber int, commitID string, path string, line int, side string, commentBody string) (*github.PullRequestComment, error)
+	CreateMultilinePRComment(ctx context.Context, owner string, repo string, pullNumber int, commitID string, path string, startLine int, endLine int, side string, commentBody string) (*github.PullRequestComment, error)
+	CreateFilePRComment(ctx context.Context, owner string, repo string, pullNumber int, commitID string, path string, commentBody string) (*github.PullRequestComment, error)
+	CreateRegularPRComment(ctx context.Context, owner string, repo string, pullNumber int, commentBody string) (*github.IssueComment, error)
+}
