+++ backend/internal/handlers/github/service.go
@@ -1,15 +1,15 @@
 package github
 
 import (
-	githubapp "github.com/CamPlume1/khoury-classroom/internal/github/app"
+	"github.com/CamPlume1/khoury-classroom/internal/github"
 	"github.com/CamPlume1/khoury-classroom/internal/storage"
 )
 
 type Service struct {
 	store  storage.Storage
-	github githubapp.GitHubApp
+	github github.GitHubApp
 }
 
-func newService(store storage.Storage, github githubapp.GitHubApp) *Service {
+func newService(store storage.Storage, github github.GitHubApp) *Service {
 	return &Service{store: store, github: github}
 }
